moves functions for pose to transforms module
