+++ source/isaaclab/isaaclab/sim/utils/__init__.py
@@ -10,4 +10,4 @@ from .prims import *  # noqa: F401, F403
 from .queries import *  # noqa: F401, F403
 from .semantics import *  # noqa: F401, F403
 from .stage import *  # noqa: F401, F403
-from .xform import *  # noqa: F401, F403
+from .transforms import *  # noqa: F401, F403
